+++ CHANGES.md
@@ -3,6 +3,7 @@
 ### 0.12.0
 
   - [BC-BREAK] Factories are given an instance of delegate instead it's own.
+  - [BC-BREAK] Factory call to make without parameters gets service from it's own instead of delegate.
 
 ### 0.11.0